in PluginTL the reference curve can be draw with a decimal step in Plugin14C an d in PluginAM the reference curve is only define with a step legal 1 In Project::load fixe a bug controle existing file before load, useful with automatic loading in the preference and if the file was deleting or moving. Deleting file old .dat when	 a model is modify without rerun Clear model and and disable access to resultview when a event was delete and when a date was delete and when a phase was delete. Change step calib resolution the range is set to (0.01 to 10000) with 2 decimal

+++ src/plugins/plugin_tl/PluginTLRefView.cpp
@@ -48,7 +48,7 @@ void PluginTLRefView::setDate(const Date& d, const ProjectSettings& settings)
         curve.mPen.setColor(Qt::blue);
         curve.mIsHisto = false;
         
-        for(int t=mSettings.mTmin; t<=mSettings.mTmax; t+=mSettings.mStep)
+        for(double t=mSettings.mTmin; t<=mSettings.mTmax; t+=mSettings.mStep)
             curve.mData[t] = ref_year - t;
         mGraph->addCurve(curve);
